Big refactoring!
* Rewritten most of the GUI and event handling code.
* Merged `conn_t`, `req_t`, and `res_t` into a single
  `piper_t` struct.
* Put repeated patterns into utility functions to make
  the code DRY-er.
* Rewritten the build system.
* Made some components more self-contained and refactored
  the whole "state machine" thing.
* Ares and PCGI are now completely independent of Rover
  except for the common shared functions, including the
  event loop.
* Added (good ol') CGI-like environment to PCGI.
* Made lots of misc changes, optimizations, and bugfixes.

Basically, restructured the whole thing.

+++ src/common/dynarr.c
@@ -34,7 +34,7 @@ int dynarr_add(dynarr_t* arr, void* ptr)
 
 void* dynarr_get(dynarr_t* arr, size_t i)
 {
-	if (i >= arr->len)
+	if ((ssize_t)i >= arr->len)
 	{
 		return NULL;
 	}
